@@ -1,8 +1,8 @@
 /*
-*
-* ListRow
-*
-*/
+ *
+ * ListRow
+ *
+ */
 
 import React from 'react';
 import { map, size } from 'lodash';
@@ -12,12 +12,18 @@ import PropTypes from 'prop-types';
 
 import styles from './styles.scss';
 
-function ListRow({ children, items, onClick}) {
+function ListRow({ children, items, onClick, ...rest }) {
   const colNumber = size(items) !== 0 ? Math.floor(12 / size(items)) : 12;
-  const data = children || map(items, (value, key) => <div className={value.className || `col-md-${colNumber}`} key={key}>{value.item}</div>);
+  const data =
+    children ||
+    map(items, (value, key) => (
+      <div className={value.className || `col-md-${colNumber}`} key={key}>
+        {value.item}
+      </div>
+    ));
 
   return (
-    <li className={styles.li}>
+    <li className={styles.li} {...rest}>
       <div className={cn(styles.container, 'row')} onClick={onClick} role="button">
         {data}
       </div>
@@ -32,8 +38,12 @@ ListRow.propTypes = {
 };
 
 ListRow.defaultProps = {
-  children: <div><FormattedMessage id="components.ListRow.empty" /></div>,
-  items: [{ item: ''}],
+  children: (
+    <div>
+      <FormattedMessage id="components.ListRow.empty" />
+    </div>
+  ),
+  items: [{ item: '' }],
   onClick: () => {},
 };
 
