@@ -6,12 +6,22 @@
 
 import React from 'react';
 import PropTypes from 'prop-types';
+import { FormattedMessage } from 'react-intl';
 
 import styles from './styles.scss';
 
 const LoadingIndicatorPage = (props) => {
   if (props.error) {
-    return <div>An error occurred</div>;
+    return (
+      <div style={{ padding: 40 }}>
+        <h2><FormattedMessage id="components.ErrorBoundary.title" /></h2>
+        <p>{props.error && props.error.toString()}</p>
+        <br />
+        <details style={{ whiteSpace: 'pre-wrap' }}>
+          {props.error.stack}
+        </details>
+      </div>
+    );
   }
 
   return (
