@@ -4,9 +4,40 @@
  *
  */
 
-export const DELETE_CONTENT_TYPE = 'ContentTypeBuilder/App/DELETE_CONTENT_TYPE';
-export const MODELS_FETCH = 'ContentTypeBuilder/App/MODELS_FETCH';
-export const MODELS_FETCH_SUCCEEDED = 'ContentTypeBuilder/App/MODELS_FETCH_SUCCEEDED';
-export const STORE_TEMPORARY_MENU = 'ContentTypeBuilder/App/STORE_TEMPORARY_MENU';
-export const TEMPORARY_CONTENT_TYPE_FIELDS_UPDATED = 'ContentTypeBuilder/App/TEMPORARY_CONTENT_TYPE_FIELDS_UPDATED';
-export const TEMPORARY_CONTENT_TYPE_POSTED = 'ContentTypeBuilder/App/TEMPORARY_CONTENT_TYPE_POSTED';
+export const ADD_ATTRIBUTE_RELATION = 'ContentTypeBuilder/App/ADD_ATTRIBUTE_RELATION';
+export const ADD_ATTRIBUTE_TO_EXISITING_CONTENT_TYPE =
+  'ContentTypeBuilder/App/ADD_ATTRIBUTE_TO_EXISITING_CONTENT_TYPE';
+export const ADD_ATTRIBUTE_TO_TEMP_CONTENT_TYPE = 'ContentTypeBuilder/App/ADD_ATTRIBUTE_TO_TEMP_CONTENT_TYPE';
+export const CANCEL_NEW_CONTENT_TYPE = 'ContentTypeBuilder/App/CANCEL_NEW_CONTENT_TYPE';
+export const CLEAR_TEMPORARY_ATTRIBUTE = 'ContentTypeBuilder/App/CLEAR_TEMPORARY_ATTRIBUTE';
+export const CLEAR_TEMPORARY_ATTRIBUTE_RELATION = 'ContentTypeBuilder/App/CLEAR_TEMPORARY_ATTRIBUTE_RELATION';
+export const CREATE_TEMP_CONTENT_TYPE = 'ContentTypeBuilder/App/CREATE_TEMP_CONTENT_TYPE';
+export const DELETE_MODEL = 'ContentTypeBuilder/App/DELETE_MODEL';
+export const DELETE_MODEL_ATTRIBUTE = 'ContentTypeBuilder/App/DELETE_MODEL_ATTRIBUTE';
+export const DELETE_MODEL_SUCCEEDED = 'ContentTypeBuilder/App/DELETE_MODEL_SUCCEEDED';
+export const DELETE_TEMPORARY_MODEL = 'ContentTypeBuilder/App/DELETE_TEMPORARY_MODEL';
+export const GET_DATA = 'ContentTypeBuilder/App/GET_DATA';
+export const GET_DATA_SUCCEEDED = 'ContentTypeBuilder/App/GET_DATA_SUCCEEDED';
+export const ON_CHANGE_EXISTING_CONTENT_TYPE_MAIN_INFOS =
+  'ContentTypeBuilder/App/ON_CHANGE_EXISTING_CONTENT_TYPE_MAIN_INFOS';
+export const ON_CHANGE_NEW_CONTENT_TYPE_MAIN_INFOS =
+  'ContentTypeBuilder/App/ON_CHANGE_NEW_CONTENT_TYPE_MAIN_INFOS';
+export const ON_CHANGE_ATTRIBUTE = 'ContentTypeBuilder/App/ON_CHANGE_ATTRIBUTE';
+export const ON_CHANGE_RELATION = 'ContentTypeBuilder/App/ON_CHANGE_RELATION';
+export const ON_CHANGE_RELATION_NATURE = 'ContentTypeBuilder/App/ON_CHANGE_RELATION_NATURE';
+export const ON_CHANGE_RELATION_TARGET = 'ContentTypeBuilder/App/ON_CHANGE_RELATION_TARGET';
+export const RESET_NEW_CONTENT_TYPE_MAIN_INFOS = 'ContentTypeBuilder/App/RESET_NEW_CONTENT_TYPE_MAIN_INFOS';
+export const RESET_EDIT_EXISTING_CONTENT_TYPE = 'ContentTypeBuilder/App/RESET_EDIT_EXISTING_CONTENT_TYPE';
+export const RESET_EXISTING_CONTENT_TYPE_MAIN_INFOS =
+  'ContentTypeBuilder/App/RESET_EXISTING_CONTENT_TYPE_MAIN_INFOS';
+export const RESET_EDIT_TEMP_CONTENT_TYPE = 'ContentTypeBuilder/App/RESET_EDIT_TEMP_CONTENT_TYPE';
+export const RESET_PROPS = 'ContentTypeBuilder/App/RESET_PROPS';
+export const SAVE_EDITED_ATTRIBUTE = 'ContentTypeBuilder/App/SAVE_EDITED_ATTRIBUTE';
+export const SAVE_EDITED_ATTRIBUTE_RELATION = 'ContentTypeBuilder/App/SAVE_EDITED_ATTRIBUTE_RELATION';
+export const SET_TEMPORARY_ATTRIBUTE = 'ContentTypeBuilder/App/SET_TEMPORARY_ATTRIBUTE';
+export const SET_TEMPORARY_ATTRIBUTE_RELATION = 'ContentTypeBuilder/App/SET_TEMPORARY_ATTRIBUTE_RELATION';
+export const SUBMIT_CONTENT_TYPE = 'ContentTypeBuilder/App/SUBMIT_CONTENT_TYPE';
+export const SUBMIT_CONTENT_TYPE_SUCCEEDED = 'ContentTypeBuilder/App/SUBMIT_CONTENT_TYPE_SUCCEEDED';
+export const SUBMIT_TEMP_CONTENT_TYPE = 'ContentTypeBuilder/App/SUBMIT_TEMP_CONTENT_TYPE';
+export const SUBMIT_TEMP_CONTENT_TYPE_SUCCEEDED = 'ContentTypeBuilder/App/SUBMIT_TEMP_CONTENT_TYPE_SUCCEEDED';
+export const UPDATE_TEMP_CONTENT_TYPE = 'ContentTypeBuilder/App/UPDATE_TEMP_CONTENT_TYPE';
