@@ -2,46 +2,31 @@ import { createSelector } from 'reselect';
 import pluginId from '../../pluginId';
 
 /**
- * Direct selector to the list state domain
+ * Direct selector to the app state domain
+ */
+const selectAppDomain = () => (state) => state.get(`${pluginId}_app`);
+
+/**
+ * Other specific selectors
  */
 
-const selectGlobalDomain = () => state => state.get(`${pluginId}_global`);
 
-const makeSelectLoading = () => createSelector(
-  selectGlobalDomain(),
-  (globalSate) => globalSate.get('loading'),
+/**
+ * Default selector used by App
+ */
+
+const makeSelectApp = () => createSelector(
+  selectAppDomain(),
+  (substate) => substate.toJS()
 );
 
-const makeSelectModels = () => createSelector(
-  selectGlobalDomain(),
-  (globalSate) => globalSate.get('models').toJS(),
+const makeSelectConnections = () => createSelector(
+  selectAppDomain(),
+  substate => substate.get('connections').toJS(),
 );
 
-const makeSelectMenu = () => createSelector(
-  selectGlobalDomain(),
-  (globalSate) => globalSate.get('menu').toJS(),
-);
-
-const selectLocationState = () => {
-  let prevRoutingState;
-  let prevRoutingStateJS;
-
-  return state => {
-    const routingState = state.get('route'); // or state.route
-
-    if (!routingState.equals(prevRoutingState)) {
-      prevRoutingState = routingState;
-      prevRoutingStateJS = routingState.toJS();
-    }
-
-    return prevRoutingStateJS;
-  };
-};
-
-
+export default makeSelectApp;
 export {
-  selectLocationState,
-  makeSelectLoading,
-  makeSelectMenu,
-  makeSelectModels,
+  selectAppDomain,
+  makeSelectConnections,
 };
