@@ -7,8 +7,8 @@ const changedFiles = listChangedFiles();
 const { take, template } = require('lodash');
 
 const cmdEslint = template(
-  'node ../../node_modules/strapi-lint/node_modules/.bin/eslint --ignore-path .gitignore --ignore-pattern "${ignore}"'
-  + ' --config ../../node_modules/strapi-lint/lib/internals/eslint/${conf}/.eslintrc.json ${params}'
+  'node ../../node_modules/strapi-lint/node_modules/.bin/eslint --ignore-path .gitignore --ignore-pattern "${ignore}"' +
+    ' --config ../../node_modules/strapi-lint/lib/internals/eslint/${conf}/.eslintrc.json ${params}',
 );
 
 const cmdFront = cmdEslint({ ignore: '/admin/build/', conf: 'front', params: 'admin' });
@@ -33,6 +33,8 @@ const watcher = (label, pckgName) => {
 const except = [
   'docs',
   'jest.config.js',
+  'jest.config.front.js',
+  'fileTransformer.js',
   'jest.config.e2e.js',
   'scripts',
   'strapi-lint',
@@ -51,7 +53,6 @@ const changedDirs = [...changedFiles]
     return take(directoryArray, toTake).join('/');
   });
 
-[...new Set(changedDirs)]
-  .forEach(directory => {
-    watcher(`Testing ${directory}`, directory);
-  });
+[...new Set(changedDirs)].forEach(directory => {
+  watcher(`Testing ${directory}`, directory);
+});
